backend/requirements.txt
agno
fastapi[standard]
openai
python-dotenv
uvicorn
sqlalchemy
aiosqlite
